Add support for plain tex
The previous configuration assumed all .tex source
files were latex and required manual configuration
to write plain tex. The new setup tries to guess
based on the contents of the file and, if
ambiguous, asks when the source is first compiled.

TeX-PDF-mode is disabled in plain-tex, which
interferes with its ability to compile.

I've changed the indent level back to 2 characters
which appears to be the level used by Knuth.