Fix company completions
Only one company completion mode can be used at a
time. However, multiple can be grouped such that
their completions are merged. Finally, when
multiple completions are specified but ungrouped,
the first completion framework in the list takes
precedence. The frameworks are set by assigning
them in a list format to `company-backends'.

To merge completion frameworks, place them in a
sublist. For instance if `company-backends' was
`((fmwk1 fmwk2))', the completions from frameworks
1 and 2 would be merged would be the sole
completions used. By contrast, if
`company-backends' was `(fmwk1 (fmwk2 fmwk3))',
then the completions from frameworks 2 and 3 would
be grouped, but would likely be ignored in favor
of framework 1.

The solution to the problem of ignored backends
used here is to set default backends that are used
in all buffers and then to set the variable
locally in specific major modes based on the
desired completions. Make sure that the format is
always `((fmwk1 fmwk2 ...))' where the ellipsis
does not contain any sublists. For an example of
how to set `company-backends' locally, see the
commit diff.