Disable gdb many windows
The many windows configuration takes up a lot of
screen real estate and I spend 90% of my time
only looking at the source window and GDB window.
It's also easy to get the information present in
the other windows so I don't see the need to have
it there all the time.

I may bring back some customized variant of this
in the future that makes more sense for my use
case.