Only enable company capf in gdb mode
The other completion backends don't provide any
useful functionality in GDB and just get in the
way. company-capf, however, is very useful. It
provides a company-backend to the
complete-at-point functionality which would
otherwise require an active TAB keypress to see.
Since GDB already provides all the necessary
completions, this is all that's needed.