Get rid of company box
I'm tentatively getting rid of company-box.
Although the documentation is nice and I at least
like the idea of different colors for different
backends and different icons for different types
of completions, I think it looks slightly worse
than the alternative and it's easy enough to get
documentation by just pressing C-h when the
completion shows. Lastly, I'm trying to be more
spartan about which backends I use anyway (I'm
starting to find that only a few are useful in
each context and the rest provide unproductive
clutter), so I should eliminate the ones I don't
use and so there isn't that much of a need to see
where it comes from.

It might be useful to bring back company box in
the future if I customize the color schemes.