Integrate evil keybindings with all modes
Broadly, I've configured evil-collection and
evil-magit. evil-magit integrates evil with magit
and evil-collection integrates evil
globally. Correspondingly, I've removed several
configurations made to use emacs when evil behaved
poorly in a certain mode. In term-mode I've
removed configurations that are no longer
necessary because of the behavior of
evil-collection there. In particular,
evil-insert-mode works well with term-char-mode
and evil-normal-mode works well with
term-line-mode so `i` and `ESC` now replace `C-c
C-j` and `C-c C-k`. It has the dual added benefit
of being a more intuitive keybinding and faster to
type.