org-layer: remove most of latex preview features
This makes the org inline latex preview facility much less
featureful. However, it should also make it much faster since the
number of packages that need to be loaded is greatly reduced and
pdflatex is used instead of luatex. This is a good tradeoff because
complex latex code should be implemented in source blocks anyway. This
will break some old code, like many inline latex snippets in the
wiki. This is acceptable because these should be migrated anyway.

If inline latex that should work stops working, this commit is almost
definitely the culprit.